Full compliance with thingspeak.com

+++ RF24L01Read.cpp
@@ -27,7 +27,7 @@ typedef struct {
   // id of the device -> max is 32 characters
   char id[16];
   // type of the device -> a simple int
-  int8_t type;
+  int16_t type;
   float val[8];
 } Payload;
 Payload p;
